Improve deploying not as admin
We don't use password input and custom credential anymore. Instead, we
run as admin and first update the owner information of the extracted
package to a non-admin user. Then we use robocopy/xcopy O param to make
them populate the owner information while copying. Finally we also
update the owner information for the symlinks/junctions we create. As a
result, all the artifacts we deploy bear out, non-admin user ownership.